requirements.txt
google-cloud-bigquery>=3.11.4
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
gspread>=5.10.0
pandas>=2.0.3
requests>=2.31.0
requests-oauthlib>=1.3.1
python-dateutil>=2.8.2
db-dtypes>=1.1.1
pyarrow>=12.0.0